+++ src/types.ts
@@ -8,7 +8,6 @@ export type Config = {
   bindAddress: string;
   fractalServerUrl: string;
   basePath: string;
-  vizarrStaticFilesPath: string;
   authorizationScheme: AuthorizationScheme;
   cacheExpirationTime: number;
   testingUsername: string | null;